feat(bezier-icons): fix to javascript because of typescript file load fail
