+++ packages/bezier-icons/scripts/generate-icon-files.js
@@ -2,7 +2,7 @@ const path = require('path')
 
 const { generateIconFiles } = require('./utils/generateIconFilesFromJson')
 
-const bezierIconsDirectory = path.resolve(__dirname, '..')
-const iconsJson = path.resolve(bezierIconsDirectory, 'icons.json')
+const bezierIconsDirectory = path.resolve(__dirname, '../icons')
+const iconsJson = path.resolve(__dirname, '../icons.json')
 
 generateIconFiles(iconsJson, bezierIconsDirectory)
